+++ packages/core/echo/echo-db/src/testing/echo-test-builder.ts
@@ -118,6 +118,7 @@ export class EchoTestPeer extends Resource {
     const db = client.constructDatabase({ spaceKey });
     await db.setSpaceRoot(rootUrl);
     await db.open();
+    this._ctx.onDispose(() => db.close());
     return db;
   }
 }
