+++ packages/gravity/blade-runner/src/redis/index.ts
@@ -2,6 +2,7 @@
 // Copyright 2023 DXOS.org
 //
 
+export * from './defaults';
 export * from './websocket-connector';
 export * from './websocket-redis-proxy';
 export * from './util';
