+++ packages/gravity/blade-runner/src/redis/util.ts
@@ -9,8 +9,6 @@ import { type Any } from '@dxos/codec-protobuf';
 import { log } from '@dxos/log';
 import { type RpcPort } from '@dxos/rpc';
 
-export const REDIS_PORT = 6379;
-
 export const createRedisRpcPort = ({
   sendClient,
   receiveClient,
